+++ models/emod-abm/README.md
@@ -1,8 +1,8 @@
-# Agent based disease transmission benchmark
+# Agent based disease transmission model
 
 ## Overview
 
-In this benchmark we run [EMOD](https://docs.idmod.org/projects/emod-generic/en/latest/index.html), a stochastic agent based disease transmission model, to look at how the reproductive number, [$R_0$](https://en.wikipedia.org/wiki/Basic_reproduction_number), and the correlation between an individual's acquisition and transmission correlation can affect the ultimate attack fraction. It turns out that this correlation can greatly affect that size of the outbreak for a fixed reproductive number. 
+This benchmark model simulates the transmission of disease in a heterogenous population using [EMOD](https://docs.idmod.org/projects/emod-generic/en/latest/index.html), a stochastic agent based disease transmission model.
 
 
 ## Authors
@@ -11,25 +11,26 @@ In this benchmark we run [EMOD](https://docs.idmod.org/projects/emod-generic/en/
 
 ## Run
 ```
-docker run -it -p 4242:4242 linusseelinger/benchmark-abm-attack-fraction:latest
+docker run -it -p 4242:4242 linusseelinger/model-emod-abm:latest
 ```
 
 ## Properties
 
 Model | Description
 ---|---
-posterior | Posterior density
+forward | Forward model
 
-### posterior
+### forward
 Mapping | Dimensions | Description
 ---|---|---
-input | [3] | [ $R_0$, variance of $R_0$, correlation between acquisition and transmission]
-output | [1] | Log posterior density
+input | [3] | [ $R_0$, variance of $R_0$, correlation between acquisition and transmission ]
+output | [1] | Dictionary with daily timeseries of 'New Infections', 'Infected', 'Infectious Population', 'Susceptible Population', 'Symptomatic Population', 'Recovered Population', and 'Exposed Population'.
+
 
 Feature | Supported
 ---|---
 Evaluate | True
-Gradient | True (via finite difference)
+Gradient | False
 ApplyJacobian | False
 ApplyHessian | False
 
@@ -47,7 +48,7 @@ None |
 
 ## Source code
 
-[Model sources here.](https://github.com/UM-Bridge/benchmarks/tree/main/benchmarks/emod-abm)
+[Model sources here.](https://github.com/UM-Bridge/benchmarks/tree/main/benchmarks/abm-attack-fraction)
 
 ## Description
 
@@ -56,10 +57,10 @@ $ \mu_{I_t} = \log\left(\frac{{R_0}}{\mu_{P}}\right) - 0.5\sigma_{I_t}^2 $
 where 
 $ \sigma_{I_t} = \log\left(\frac{\sigma_{R_0}^2}{2R_0^2} + 0.5\right). $
 
-One of the challenges associated with this benchmark is that it is stochastic.  Even for the same parameters the total number of infections will differ between simulations. Furthermore, it is not guaranteed that there will always be an outbreak: sometimes, by chance, there are not enough initial infections to generate a large outbreak.  You can reduce this challenge by increasing the number of initial infections via the `daily_import_pressures` configuration parameter (e.g., from 1 infections per day to 10 infections per day for the first 5 days).
+One of the challenges associated with this model is that it is stochastic.  Even for the same parameters the total number of infections will differ between simulations. Furthermore, it is not guaranteed that there will always be an outbreak: sometimes, by chance, there are not enough initial infections to generate a large outbreak.  You can reduce this challenge by increasing the number of initial infections via the `daily_import_pressures` configuration parameter (e.g., from 1 infections per day to 10 infections per day for the first 5 days).
 
 Furthermore, the likelihood of an individual to acquire and then transmit the disease is correlated. There is no waning immunity. The benchmark is fitting to an attack fraction of 0.40 with a standard deviation of 0.025. There are bounds on $R_0$ > 0; the variance of $R_0$ > 0; and the correlation between acquisition and transmission must lie between 0 and 1 (inclusive).
 
-The simulation terminates when there are no longer any infected individuals (minimum run time of 50 days) and assesses the final attack fraction in the population (the fraction of people who were infected over the course of the disease outbreak).
+The simulation terminates when there are no longer any infected individuals (minimum run time of 50 days) and return daily timeseries 'New Infections', 'Infected', 'Infectious Population', 'Susceptible Population', 'Symptomatic Population', 'Recovered Population', and 'Exposed Population'.
